BUILD-04 | split on header and source
