+++ CMakeLists.txt
@@ -25,6 +25,8 @@ project(meta_print_ip VERSION ${MAJOR_VERSION}.${MINOR_VERSION}.${PATCH_VERSION}
 
 add_executable(${PROJECT_NAME} ${PROJECT_SOURCE_DIR}/src/main.cpp)
 
+target_include_directories(${PROJECT_NAME} PUBLIC ${PROJECT_SOURCE_DIR}/include)
+
 set(CPACK_GENERATOR "DEB")
 set(CPACK_DEBIAN_PACKAGE_MAINTAINER "KK")
 set(CPACK_DEBIAN_PACKAGE_SHLIBDEPS ON)
